+++ src/graphql/entity/create_field/resolve_fields/resolve_document_field/mod.rs
@@ -1,6 +1,6 @@
 use async_graphql::{indexmap::IndexMap, Name, Value};
 use bson::{Bson, Document};
-use log::debug;
+use log::{debug, trace};
 
 use crate::{
     configuration::subgraph::entities::service_entity_field::ServiceEntityFieldConfig,
@@ -11,7 +11,7 @@ impl ServiceEntity {
     pub fn resolve_document_object_id_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving Object ID Scalar");
         let resolved =
             field
@@ -19,9 +19,15 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::ObjectID(object_id) => Ok(Value::from(object_id.to_string())),
-            DocumentValue::Null => Ok(Value::Null),
-            DocumentValue::None => Ok(Value::Null),
+            DocumentValue::ObjectID(object_id) => Ok(Some(Value::from(object_id.to_string()))),
+            DocumentValue::ObjectIDArray(object_ids) => Ok(Some(Value::List(
+                object_ids
+                    .into_iter()
+                    .map(|object_id| Value::from(object_id.to_string()))
+                    .collect(),
+            ))),
+            DocumentValue::Null => Ok(Some(Value::Null)),
+            DocumentValue::None => Ok(None),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for object id scalar",
             )),
@@ -31,7 +37,7 @@ impl ServiceEntity {
     pub fn resolve_document_string_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving String Scalar");
         let resolved =
             field
@@ -39,12 +45,12 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::String(value) => Ok(Value::from(value)),
-            DocumentValue::StringArray(values) => Ok(Value::List(
+            DocumentValue::String(value) => Ok(Some(Value::from(value))),
+            DocumentValue::StringArray(values) => Ok(Some(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
-            )),
-            DocumentValue::None => Ok(Value::Null),
-            DocumentValue::Null => Ok(Value::Null),
+            ))),
+            DocumentValue::None => Ok(None),
+            DocumentValue::Null => Ok(Some(Value::Null)),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for string scalar",
             )),
@@ -54,7 +60,7 @@ impl ServiceEntity {
     pub fn resolve_document_int_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving Int Scalar");
 
         let resolved =
@@ -63,12 +69,12 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::Int(value) => Ok(Value::from(value)),
-            DocumentValue::IntArray(values) => Ok(Value::List(
+            DocumentValue::Int(value) => Ok(Some(Value::from(value))),
+            DocumentValue::IntArray(values) => Ok(Some(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
-            )),
-            DocumentValue::None => Ok(Value::Null),
-            DocumentValue::Null => Ok(Value::Null),
+            ))),
+            DocumentValue::None => Ok(None),
+            DocumentValue::Null => Ok(Some(Value::Null)),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for int scalar",
             )),
@@ -78,7 +84,7 @@ impl ServiceEntity {
     pub fn resolve_document_boolean_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving Boolean Scalar");
 
         let resolved =
@@ -87,12 +93,12 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::Boolean(value) => Ok(Value::from(value)),
-            DocumentValue::BooleanArray(values) => Ok(Value::List(
+            DocumentValue::Boolean(value) => Ok(Some(Value::from(value))),
+            DocumentValue::BooleanArray(values) => Ok(Some(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
-            )),
-            DocumentValue::Null => Ok(Value::Null),
-            DocumentValue::None => Ok(Value::Null),
+            ))),
+            DocumentValue::Null => Ok(Some(Value::Null)),
+            DocumentValue::None => Ok(None),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for boolean scalar",
             )),
@@ -102,7 +108,7 @@ impl ServiceEntity {
     pub fn resolve_document_uuid_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving UUID Scalar");
 
         let resolved =
@@ -111,15 +117,15 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::UUID(value) => Ok(Value::from(value.to_string())),
-            DocumentValue::UUIDArray(values) => Ok(Value::List(
+            DocumentValue::UUID(value) => Ok(Some(Value::from(value.to_string()))),
+            DocumentValue::UUIDArray(values) => Ok(Some(Value::List(
                 values
                     .into_iter()
                     .map(|value| Value::from(value.to_string()))
                     .collect(),
-            )),
-            DocumentValue::Null => Ok(Value::Null),
-            DocumentValue::None => Ok(Value::Null),
+            ))),
+            DocumentValue::Null => Ok(Some(Value::Null)),
+            DocumentValue::None => Ok(None),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for UUID scalar",
             )),
@@ -129,7 +135,7 @@ impl ServiceEntity {
     pub fn resolve_document_datetime_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving DateTime Scalar");
 
         let resolved =
@@ -138,15 +144,15 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::DateTime(value) => Ok(Value::from(value.to_rfc3339())),
-            DocumentValue::DateTimeArray(values) => Ok(Value::List(
+            DocumentValue::DateTime(value) => Ok(Some(Value::from(value.to_rfc3339()))),
+            DocumentValue::DateTimeArray(values) => Ok(Some(Value::List(
                 values
                     .into_iter()
                     .map(|value| Value::from(value.to_rfc3339()))
                     .collect(),
-            )),
-            DocumentValue::Null => Ok(Value::Null),
-            DocumentValue::None => Ok(Value::Null),
+            ))),
+            DocumentValue::Null => Ok(Some(Value::Null)),
+            DocumentValue::None => Ok(None),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for DateTime scalar",
             )),
@@ -166,7 +172,7 @@ impl ServiceEntity {
 
         for (key, bson) in document.into_iter() {
             let name = Name::new(key);
-            debug!("---Found BSON Element Type: {:?}", bson.element_type());
+            trace!("Found BSON Element Type: {:?}", bson.element_type());
 
             let field = ServiceEntityFieldConfig::get_field(
                 field.fields.clone().unwrap_or(Vec::new()),
@@ -176,7 +182,9 @@ impl ServiceEntity {
                 .scalar
                 .clone()
                 .document_field_to_async_graphql_value(document, &field)?;
-            index_map.insert(name, value);
+            if let Some(value) = value {
+                index_map.insert(name, value);
+            }
         }
         Ok(index_map)
     }
@@ -184,12 +192,13 @@ impl ServiceEntity {
     pub fn resolve_document_object_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
-        debug!("Resolving Object Scalar");
+    ) -> Result<Option<Value>, async_graphql::Error> {
+        debug!("Resolve Document Object Scalar");
         let value = document.get(&field.name);
 
         if value.is_none() {
-            return Ok(Value::Null);
+            trace!("Value is None, returning null");
+            return Ok(None);
         }
 
         if field.list.unwrap_or(false) {
@@ -200,21 +209,21 @@ impl ServiceEntity {
                     values.push(Value::Object(value));
                 }
 
-                return Ok(Value::List(values));
+                return Ok(Some(Value::List(values)));
             } else {
-                return Ok(Value::List(vec![]));
+                return Ok(Some(Value::List(vec![])));
             }
         }
 
         let index_map = ServiceEntity::parse_bson_object(value.unwrap(), field)?;
 
-        Ok(Value::from(index_map))
+        Ok(Some(Value::from(index_map)))
     }
 
     pub fn resolve_document_enum_scalar(
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving Enum Scalar");
 
         let resolved =
@@ -223,12 +232,12 @@ impl ServiceEntity {
                 .get_from_document(document, &field.name, field.list.unwrap_or(false))?;
 
         match resolved {
-            DocumentValue::String(value) => Ok(Value::from(value)),
-            DocumentValue::StringArray(values) => Ok(Value::List(
+            DocumentValue::String(value) => Ok(Some(Value::from(value))),
+            DocumentValue::StringArray(values) => Ok(Some(Value::List(
                 values.into_iter().map(|value| Value::from(value)).collect(),
-            )),
-            DocumentValue::None => Ok(Value::Null),
-            DocumentValue::Null => Ok(Value::Null),
+            ))),
+            DocumentValue::None => Ok(None),
+            DocumentValue::Null => Ok(Some(Value::Null)),
             _ => Err(async_graphql::Error::from(
                 "Invalid result type for enum scalar",
             )),