+++ src/scalar_option/document_field_to_async_graphql_value.rs
@@ -15,7 +15,7 @@ impl ScalarOption {
         self,
         document: &Document,
         field: &ServiceEntityFieldConfig,
-    ) -> Result<Value, async_graphql::Error> {
+    ) -> Result<Option<Value>, async_graphql::Error> {
         debug!("Resolving Document Field: {:?}", field.name);
 
         let value = match self {